// src/app/Shared/models/Customer.ts
// tslint:disable-next-line:class-name
export default class M_Customer {
    CustomerId: number;
    CustomerName: string;
    MobileNo: string ;
    EmailId: string;
    Password: string;
    ConfirmPassword: string;
    ProfileImage: string;
    Active: boolean;
    CreatedDate: Date;
    UpdatedDate: Date;
    StatusCode: number;
    Status: string;
}